[Migration] Changes from 8.10.0 into 15.0.0 (#1273)
* Merge pull request #1176 from OntimizeWeb/release/8.11.0
Release 8.11.0

* Merge pull request #133 from OntimizeWeb/hotfix/8.5.2 and Merge pull request #126 from OntimizeWeb/hotfix/8.5.1

* Merged ontimize-web-ngx-theming@8.5.0 changes

* Merge pull request #1181 from OntimizeWeb/hotfix/8.11.1 Hotfix 8.11.1

* Merge pull request #1190 from OntimizeWeb/hotfix/8.11.2
Hotfix 8.11.2

* Merge pull request #1200 from OntimizeWeb/sync/release-8.11.3-to-develop

Sync release 8.11.3 to develop

* Merge pull request #1217 from OntimizeWeb/release/8.12.0
Release 8.12.0

* Merge pull request #1227 from OntimizeWeb/sync/release-8.12.1-to-develop

Sync release 8.12.1 to develop

* Merge pull request #1232 from OntimizeWeb/hotfix/8.12.2
Hotfix 8.12.2

* Merge pull request #1255 from OntimizeWeb/release/8.13.0
Release 8.13.0

* Merge pull request #1271 from OntimizeWeb/release/8.14.0
Release 8.14.0

* Updated version to 15.0.0-SNAPSHOT-1

* Remove this unused import of 'DEFAULT_OUTPUTS_O_TABLE_CELL_EDITOR'.

* 'subscribe' is deprecated. Instead of passing separate callback arguments, use an observer argument

---------

+++ projects/ontimize-web-ngx/src/lib/components/form/navigation/o-form-navigation.component.ts
@@ -249,7 +249,10 @@ export class OFormNavigationComponent implements OnDestroy {
         this.router.navigate(route, extras).then((navigationDone: boolean) => {
           if (navigationDone) {
             this.currentIndex = index;
-            this.navigationService.removeLastItem();
+            const url = this.router.routerState.snapshot.url.split('?')[0];
+            if (url !== navData.url) {
+              this.navigationService.removeLastItem();
+            }
           }
         });
       }
